normal attack by timeline

+++ Assets/Scripts/Game/Modules/Role/UploadMainRolePosSystem.cs
@@ -39,12 +39,12 @@ public class UploadMainRolePosSystem : BaseComponentSystem
             synchInfo.LastUploadPos = pos;
             synchInfos[i] = synchInfo;
             scene_walk.request walk = new scene_walk.request();
-            walk.start_x = (int)(pos.x*GameConst.RealToLogic);
-            walk.start_y = (int)(pos.y*GameConst.RealToLogic);
-            walk.start_z = (int)(pos.z*GameConst.RealToLogic);
-            walk.end_x = (int)(targetPos.x*GameConst.RealToLogic);
-            walk.end_y = (int)(targetPos.y*GameConst.RealToLogic);
-            walk.end_z = (int)(targetPos.z*GameConst.RealToLogic);
+            walk.start_x = (long)(pos.x*GameConst.RealToLogic);
+            walk.start_y = (long)(pos.y*GameConst.RealToLogic);
+            walk.start_z = (long)(pos.z*GameConst.RealToLogic);
+            walk.end_x = (long)(targetPos.x*GameConst.RealToLogic);
+            walk.end_y = (long)(targetPos.y*GameConst.RealToLogic);
+            walk.end_z = (long)(targetPos.z*GameConst.RealToLogic);
             walk.time = synchTime;
             NetMsgDispatcher.GetInstance().SendMessage<Protocol.scene_walk>(walk);
             lastSynchTime = Time.time;
